+++ changelog.md
@@ -11,6 +11,8 @@
 * **Added** - WooCommerce: Product archive disable/enable image and title links conditional logic.
 * **Added** - WooCommerce: Single product show/hide price and 'Add to Cart' button option.
 * **Added** - WooCommerce: Single product show/hide price and 'Add to Cart' button message display (optional).
+* **Added** - WooCommerce: Single product show/hide price and 'Add to Cart' button message typography settings.
+* **Added** - WooCommerce: Single product show/hide price and 'Add to Cart' button message color styling settings.
 * **Added** - WooCommerce: Single product - option to include 'My Account' link into the conditional message.
 * **Tweak** - WooCommerce: Product archive rating separated from price as an individual element.
 * **Fixed** - WooCommerce: Mixed up translation filters for Next/Prev product navigation.